Refactor testutil.NewStorage into a new package
This avoids a circular dependency between the testutil and storage
packages.
