+++ promql/test.go
@@ -28,6 +28,7 @@ import (
 
 	"github.com/prometheus/prometheus/pkg/labels"
 	"github.com/prometheus/prometheus/storage"
+	"github.com/prometheus/prometheus/util/teststorage"
 	"github.com/prometheus/prometheus/util/testutil"
 )
 
@@ -509,7 +510,7 @@ func (t *Test) clear() {
 	if t.cancelCtx != nil {
 		t.cancelCtx()
 	}
-	t.storage = testutil.NewStorage(t)
+	t.storage = teststorage.New(t)
 
 	opts := EngineOpts{
 		Logger:        nil,
@@ -623,7 +624,7 @@ func (ll *LazyLoader) clear() {
 	if ll.cancelCtx != nil {
 		ll.cancelCtx()
 	}
-	ll.storage = testutil.NewStorage(ll)
+	ll.storage = teststorage.New(ll)
 
 	opts := EngineOpts{
 		Logger:        nil,
